Refactor how all the dialogs are handled into a separate module
This required the factoring out of a card ui module as well.
Making this change simplifies the creation and management of new dialogs, and should allow doing things like having multiple dialogs open and putting the most recently interacted-with dialog on top. It should also make it simpler to have multiple dialogs of the same type open in the future, although that is not planned in the near term.